fix: give fidelity bond its own view controller, make dates dynamic, show FB as an invoice not via tx creator.
